# swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Stock Battle
  description: Stock Battle API docs.
  version: 1.0.0
...
servers:
  - url: /api/
...
paths:
...
  /user-stock:
    post:
      tags:
        - user stock
      summary: Register stock
      operationId: registerStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStockRequest'
      responses:
        200:
          description: successful operation
  /user-stock/{roomId}:
    get:
      tags:
        - user stock
      summary: The list of stocks registered by users in room
      operationId: getUserStocks
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserStockResponse'
...
components:
  schemas:
...
    UserStockRequest:
      type: object
      properties:
        roomId:
          type: integer
        ticker:
          type: string
        amount:
          type: integer
    UserStockResponse:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        ticker:
          type: string
        amount:
          type: integer
...